fix typos and some small refactorings to the api
